+++ api/src/main/java/net/automatalib/automata/vpda/OneSEVPA.java
@@ -19,8 +19,13 @@ import java.util.List;
 
 import com.google.common.collect.Iterables;
 import net.automatalib.automata.concepts.FiniteRepresentation;
+import net.automatalib.automata.concepts.InputAlphabetHolder;
 import net.automatalib.automata.concepts.SuffixOutput;
+import net.automatalib.automata.vpda.OneSEVPAGraphView.SevpaViewEdge;
+import net.automatalib.graphs.Graph;
+import net.automatalib.graphs.concepts.GraphViewable;
 import net.automatalib.ts.acceptors.DeterministicAcceptorTS;
+import net.automatalib.words.VPDAlphabet;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 /**
@@ -38,8 +43,14 @@ import org.checkerframework.checker.nullness.qual.Nullable;
  *
  * @author Malte Isberner
  */
-public interface OneSEVPA<L, I>
-        extends DeterministicAcceptorTS<State<L>, I>, SuffixOutput<I, Boolean>, FiniteRepresentation {
+public interface OneSEVPA<L, I> extends DeterministicAcceptorTS<State<L>, I>,
+                                        SuffixOutput<I, Boolean>,
+                                        InputAlphabetHolder<I>,
+                                        GraphViewable,
+                                        FiniteRepresentation {
+
+    @Override
+    VPDAlphabet<I> getInputAlphabet();
 
     int encodeStackSym(L srcLoc, I callSym);
 
@@ -80,4 +91,9 @@ public interface OneSEVPA<L, I>
     }
 
     L getInitialLocation();
+
+    @Override
+    default Graph<L, SevpaViewEdge<L, I>> graphView() {
+        return new OneSEVPAGraphView<>(this);
+    }
 }